created load-medication entrypoint created loaded-medication entry point implemented save droneLoad and get-loaded-medication controller methods created get-loaded medication method at useCase add validators to create a new drone created entity LoadedMedication to handle data controller response implemented get medication controller methods

+++ infrastructure/driven-adapters/jpa-repository/src/main/java/co/com/everluis/jpa/entities/MedicationEntity.java
@@ -20,7 +20,7 @@ public class MedicationEntity {
     @Column(name = "medication_name")
     private String medicationName;
     @Column(name = "weight")
-    private int medicationWeight;
+    private Double medicationWeight;
     @Column(name = "image_url")
     private String medicationImage;
 }
